requirements.txt
vvhgvs@git+https://github.com/openvar/vv_hgvs@master
VariantFormatter@git+https://github.com/openvar/variantFormatter@master
VariantValidator@git+https://github.com/openvar/variantValidator@master
django
celery
django-celery-results
django-celery-beat
django-widget-tweaks
django-cookie-law
django-allauth
django-recaptcha
django_countries
jwt
weasyprint
mod_wsgi
pytest-django
